fix(sql): a ninth sweep — a named window was evaluated as OVER ()
Window functions, DML/RETURNING and CTEs probed against PostgreSQL 14.13, both
sides through the same psycopg client. CTEs came back clean (23 of 23,
recursive and data-modifying included). Windows did not — four findings were
silent, each a plausible number computed from a window nobody wrote:

- A `WINDOW w AS (...)` reference lost its whole definition. sqlglot keeps the
  definition on the SELECT and leaves the reference as a bare alias with no
  partition, order or frame, which is exactly what consumers read.
- `EXCLUDE` rode on the frame spec and nothing read it, so EXCLUDE CURRENT ROW
  answered the unexcluded frame.
- `NULLS FIRST` in a window ORDER BY was ignored. Placing NULLs by direction
  alone reproduces PG's defaults exactly, so the flag looked right until
  somebody wrote it explicitly.
- Window `sum` / `avg` used their own typing rules where the GROUP BY path had
  long since got it right: sum(int4) declared int4, avg declared float8 and
  divided as a float.

Also: `agg(...) FILTER (WHERE ...) OVER (...)` was rejected with 42803 naming
an ordinary column, because the FILTER node sits between the aggregate and its
Window and the guard looked one level up; a subquery in RETURNING crashed with
XX000 (catalog hardcoded to None); ntile typed as int8 where PG says int4.

Added string_agg / array_agg as window functions and `UPDATE SET (a,b)=(x,y)`
in all three row-constructor spellings.

Two existing tests asserted the old behaviour and are updated to PostgreSQL's
answers — `test_avg_over_numeric` said so in its own comment ("The SCALE still
differs from PG"), which this fixes. tests/test_sql_sweep_nine.py pins the
rest; four items measured but not fixed are in tasks/backlog.md.

+++ tests/test_sql_window_frames.py
@@ -49,7 +49,9 @@ def test_ntile_two_buckets(storage, session):
     # 5 rows / 2 buckets: the first bucket gets the extra row (3 + 2).
     res = q(storage, session, "SELECT id, NTILE(2) OVER (ORDER BY id) AS nt FROM t ORDER BY id")
     assert res.rows == [(1, 1), (2, 1), (3, 1), (4, 2), (5, 2)]
-    assert res.columns[1].type_tag == "int8"
+    # `ntile` is the one integer window function Postgres types as int4, not
+    # bigint (`pg_typeof(ntile(2) OVER ...)` is `integer`, measured 14.13).
+    assert res.columns[1].type_tag == "int4"
 
 
 def test_ntile_per_partition(storage, session):
